feat: refactor Itineraries related tests into a directory
